Edited the 100-prime_factors.c file

+++ 0x04-more_functions_nested_loops/100-prime_factors_mysol.c
@@ -8,7 +8,7 @@
 int main(void)
 {
 	int i = 2;
-	long long int num = 612852475143;
+	unsigned long num = 612852475143;
 
 	while (1)
 	{
